Fixed bug #1610 (Eliminates 15 warnings of some types in modules of "clients" catalog).

+++ clients/mwm/WmMenu.c
@@ -848,8 +848,8 @@ static MenuItem *MakeMenuItemFromTemplate (MenuItem *template, String name,
 static MenuItem *MakeClientCommandMenuItem (String label, String funcArgs)
 {
     return(MakeMenuItem(label, F_InvokeCommand, funcArgs,
-			(KeySym) NULL, (unsigned int)0,
-			(KeyCode) NULL, (String)NULL));
+			(KeySym)(unsigned)NULL, (unsigned int)0,
+			(KeyCode)(unsigned)NULL, (String)NULL));
 }
